Split CustomSetup into new file
