+++ src/utils/draftHelpers.ts
@@ -3,9 +3,10 @@
  */
 
 import { getRareWeightMultiplier } from '../constants/balancingConfig'
+import { isDraftFilteringDebugEnabled } from '../constants/gameConfig'
 import { FACTION, RARITY, TAGS, TYPE } from '../constants/types'
-import type { ItemType, Item, Player, GameConfig } from '../types'
 import { MASTER_DB } from '../data/itemsByWarbond'
+import type { DraftState, GameConfig, Item, ItemType, Player } from '../types'
 import {
     anyItemHasTag,
     getItemById,
@@ -14,7 +15,6 @@ import {
     playerHasAccessToArmorCombo,
     type ArmorCombo,
 } from './itemHelpers'
-import { isDraftFilteringDebugEnabled } from '../constants/gameConfig'
 
 /**
  * Weighted pool item for regular items
@@ -504,3 +504,15 @@ export const generateDraftHand = (
 
     return hand
 }
+// Helper to create a default DraftState with all required properties
+export const createDraftState = (overrides: Partial<DraftState> = {}): DraftState => ({
+    activePlayerIndex: 0,
+    roundCards: [],
+    isRerolling: false,
+    pendingStratagem: null,
+    extraDraftRound: 0,
+    draftOrder: [],
+    isRetrospective: false,
+    retrospectivePlayerIndex: null,
+    ...overrides,
+})
